# openapi/openapi.yaml
openapi: 3.0.1
info:
  title: BaCS API
  description: API сервиса по бронированию рабочих мест в коворкингах BaCS.
  version: v1
paths:
...
  /resources:
    get:
      tags:
        - Resources
      summary: Получить список ресурсов.
      parameters:
        - name: ids
          in: query
          description: Фильтр по ID ресурсов.
          schema:
            type: array
            items:
              type: string
              format: uuid
            description: Фильтр по ID ресурсов.
            nullable: true
        - name: locationIds
          in: query
          description: Фильтр по ID локаций.
          schema:
            type: array
            items:
              type: string
              format: uuid
            description: Фильтр по ID локаций.
            nullable: true
        - name: types
          in: query
          description: Фильтр по типам ресурсов.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ResourceType'
            description: Фильтр по типам ресурсов.
            nullable: true
        - name: offset
          in: query
          schema:
            minimum: 0
            type: integer
            format: int32
        - name: limit
          in: query
          schema:
            maximum: 100
            minimum: 0
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceDtoPaginatedResponse'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags:
        - Resources
      summary: Создать ресурс.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/resources/{resourceId}':
    get:
      tags:
        - Resources
      summary: Получить ресурс по ID.
      parameters:
        - name: resourceId
          in: path
          description: ID ресурса.
          required: true
          schema:
            type: string
            description: ID ресурса.
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceDto'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - Resources
      summary: Обновить ресурс.
      parameters:
        - name: resourceId
          in: path
          description: ID ресурса.
          required: true
          schema:
            type: string
            description: ID ресурса.
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResourceRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Resources
      summary: Удалить ресурс.
      parameters:
        - name: resourceId
          in: path
          description: ID ресурса.
          required: true
          schema:
            type: string
            description: ID ресурса.
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/resources/{resourceId}/image':
    put:
      tags:
        - Resources
      summary: Загрузить фотографию ресурса.
      parameters:
        - name: resourceId
          in: path
          description: ID ресурса.
          required: true
          schema:
            type: string
            description: ID ресурса.
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  description: Фотография ресурса.
                  format: binary
            encoding:
              file:
                style: form
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Resources
      summary: Удалить фотографию ресурса.
      parameters:
        - name: resourceId
          in: path
          description: ID ресурса.
          required: true
          schema:
            type: string
            description: ID ресурса.
            format: uuid
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /users:
    get:
      tags:
        - Users
      summary: Получить список пользователей.
      parameters:
        - name: ids
          in: query
          description: Фильтр по ID пользователей.
          schema:
            type: array
            items:
              type: string
              format: uuid
            description: Фильтр по ID пользователей.
            nullable: true
        - name: offset
          in: query
          schema:
            minimum: 0
            type: integer
            format: int32
        - name: limit
          in: query
          schema:
            maximum: 100
            minimum: 0
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDtoPaginatedResponse'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/users/{userId}':
    get:
      tags:
        - Users
      summary: Получить пользователя по ID.
      parameters:
        - name: userId
          in: path
          description: ID пользователя.
          required: true
          schema:
            type: string
            description: ID пользователя.
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - Users
      summary: Обновить пользователя.
      parameters:
        - name: userId
          in: path
          description: ID пользователя.
          required: true
          schema:
            type: string
            description: ID пользователя.
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - Users
      summary: Удалить пользователя.
      parameters:
        - name: userId
          in: path
          description: ID пользователя.
          required: true
          schema:
            type: string
            description: ID пользователя.
            format: uuid
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
...
    CreateResourceRequest:
      required:
        - locationId
        - name
        - type
      type: object
      properties:
        locationId:
          minLength: 1
          type: string
          description: 'ID локации, в которой будет создан ресурс.'
          format: uuid
        name:
          minLength: 1
          type: string
          description: Название ресурса.
        description:
          type: string
          description: Описание ресурса.
          nullable: true
        equipment:
          type: array
          items:
            type: string
          description: 'Оборудование, прикреплённое к ресурсу.'
          nullable: true
        type:
          $ref: '#/components/schemas/ResourceType'
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    ResourceDto:
      type: object
      properties:
        id:
          type: string
          description: ID ресурса.
          format: uuid
        locationId:
          type: string
          description: ID локации.
          format: uuid
        name:
          type: string
          description: Название ресурса.
          nullable: true
        description:
          type: string
          description: Описание ресурса.
          nullable: true
        floor:
          type: integer
          description: 'Этаж, на котором расположен ресурс.'
          format: int32
        equipment:
          type: array
          items:
            type: string
          description: 'Оборудование, прикреплённое к ресурсу.'
          nullable: true
        type:
          $ref: '#/components/schemas/ResourceType'
        imageUrl:
          type: string
          description: Ссылка на фотографию ресурса.
          nullable: true
      additionalProperties: false
    ResourceDtoPaginatedResponse:
      type: object
      properties:
        collection:
          type: array
          items:
            $ref: '#/components/schemas/ResourceDto'
          nullable: true
        totalCount:
          type: integer
          format: int32
      additionalProperties: false
    ResourceType:
      enum:
        - Unspecified
        - MeetingRoom
        - Workplace
      type: string
...
    UpdateResourceRequest:
      required:
        - name
        - type
      type: object
      properties:
        name:
          minLength: 1
          type: string
          description: Название ресурса.
        description:
          type: string
          description: Описание ресурса.
          nullable: true
        equipment:
          type: array
          items:
            type: string
          description: 'Оборудование, прикреплённое к ресурсу.'
          nullable: true
        type:
          $ref: '#/components/schemas/ResourceType'
      additionalProperties: false
    UpdateUserRequest:
      required:
        - email
        - name
      type: object
      properties:
        name:
          minLength: 1
          type: string
          description: ФИО пользователя.
        email:
          minLength: 1
          type: string
          description: Email-адрес пользователя.
          format: email
      additionalProperties: false
    UserDto:
      type: object
      properties:
        id:
          type: string
          description: ID пользователя.
          format: uuid
        username:
          type: string
          description: Уникальный логин пользователя.
          nullable: true
        email:
          type: string
          description: Email-адрес пользователя.
          nullable: true
        name:
          type: string
          description: ФИО пользователя.
          nullable: true
      additionalProperties: false
    UserDtoPaginatedResponse:
      type: object
      properties:
        collection:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
          nullable: true
        totalCount:
          type: integer
          format: int32
      additionalProperties: false
  securitySchemes:
    Keycloak:
      type: oauth2
      description: Аутентификация через Keycloak
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:7080/keycloak/realms/bacs/protocol/openid-connect/auth
          tokenUrl: http://localhost:7080/keycloak/realms/bacs/protocol/openid-connect/token
          scopes:
            bacs-api: 'Обязательный скоуп, необходимый для авторизации на BaCS API'
          x-usePkce: SHA-256
security:
  - Keycloak: [ ]